requirements.txt
openslide
pillow
numpy
opencv-python
os
pathlib
matplotlib
imageio
argparse
torchvision
torch
tensorflow
datetime
sys
tqdm
scikit-learn
scikit-image
simpleitk
random